# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "API for managing a mechanic shop's customers, mechanics, service tickets, and inventory."
  version: "1.0.0"
host: "127.0.0.1:5000"
basePath: "/"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
...
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "Enter the word Bearer, a space, and then the token: Bearer <token>"
paths:
...
  /tickets/:
    post:
      tags:
        - "Tickets"
      operationId: "createTicket"
      summary: "Create a ticket"
      description: "Creates a service ticket for an existing customer."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a ticket."
          required: true
          schema:
            $ref: "#/definitions/TicketPayload"
      responses:
        201:
          description: "Ticket created successfully."
          schema:
            $ref: "#/definitions/TicketResponse"
        400:
          $ref: "#/responses/BadRequest"
        404:
          $ref: "#/responses/NotFound"
        429:
          $ref: "#/responses/TooManyRequests"
  /tickets/my-tickets:
    get:
      tags:
        - "Tickets"
      operationId: "getCustomerTickets"
      summary: "Get the logged-in customer's tickets"
      description: "Returns every ticket belonging to the customer identified by the customer token."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Tickets retrieved successfully."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/TicketResponse"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        429:
          $ref: "#/responses/TooManyRequests"
  /tickets/{ticket_id}:
    parameters:
      - in: "path"
        name: "ticket_id"
        description: "Unique ID of the ticket."
        required: true
        type: "integer"
        minimum: 1

    get:
      tags:
        - "Tickets"
      operationId: "getTicket"
      summary: "Get one ticket"
      description: "Returns a ticket by ID."
      responses:
        200:
          description: "Ticket retrieved successfully."
          schema:
            $ref: "#/definitions/TicketResponse"
        404:
          $ref: "#/responses/NotFound"
        429:
          $ref: "#/responses/TooManyRequests"

    put:
      tags:
        - "Tickets"
      operationId: "updateTicket"
      summary: "Update a ticket"
      description: "Updates a ticket and verifies that its customer exists."
      parameters:
        - in: "body"
          name: "body"
          description: "Complete replacement values for the ticket."
          required: true
          schema:
            $ref: "#/definitions/TicketPayload"
      responses:
        200:
          description: "Ticket updated successfully."
          schema:
            $ref: "#/definitions/TicketResponse"
        400:
          $ref: "#/responses/BadRequest"
        404:
          $ref: "#/responses/NotFound"
        429:
          $ref: "#/responses/TooManyRequests"

    delete:
      tags:
        - "Tickets"
      operationId: "deleteTicket"
      summary: "Delete a ticket"
      description: "Deletes a ticket. A valid mechanic token is required."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Ticket deleted successfully."
          schema:
            $ref: "#/definitions/MessageResponse"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        404:
          $ref: "#/responses/NotFound"
        429:
          $ref: "#/responses/TooManyRequests"
  /tickets/{ticket_id}/edit:
    parameters:
      - in: "path"
        name: "ticket_id"
        description: "Unique ID of the ticket whose mechanics will be edited."
        required: true
        type: "integer"
        minimum: 1

    put:
      tags:
        - "Tickets"
      operationId: "editTicketMechanics"
      summary: "Add or remove mechanics from a ticket"
      description: "Adds mechanics listed in add_ids and removes mechanics listed in remove_ids. Use an empty array when no mechanics need to be added or removed."
      parameters:
        - in: "body"
          name: "body"
          description: "Mechanic IDs to add to or remove from the ticket."
          required: true
          schema:
            $ref: "#/definitions/EditTicketMechanicsPayload"
      responses:
        200:
          description: "Ticket mechanics updated successfully."
          schema:
            $ref: "#/definitions/TicketResponse"
        400:
          $ref: "#/responses/BadRequest"
        404:
          $ref: "#/responses/NotFound"
        429:
          $ref: "#/responses/TooManyRequests"
  /tickets/{ticket_id}/add-part/{inventory_id}:
    parameters:
      - in: "path"
        name: "ticket_id"
        description: "Unique ID of the ticket receiving the inventory part."
        required: true
        type: "integer"
        minimum: 1
      - in: "path"
        name: "inventory_id"
        description: "Unique ID of the inventory part to add."
        required: true
        type: "integer"
        minimum: 1

    put:
      tags:
        - "Tickets"
      operationId: "addPartToTicket"
      summary: "Add an inventory part to a ticket"
      description: "Adds the requested quantity of a part. If the part is already on the ticket, its quantity is increased."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Quantity to add. If omitted, quantity defaults to 1."
          required: false
          schema:
            $ref: "#/definitions/AddPartToTicketPayload"
      responses:
        200:
          description: "Inventory part added to the ticket successfully."
          schema:
            $ref: "#/definitions/TicketResponse"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        404:
          $ref: "#/responses/NotFound"
        429:
          $ref: "#/responses/TooManyRequests"
...
responses:
  BadRequest:
    description: "The request is invalid or authentication is required."
    schema:
      $ref: "#/definitions/ErrorResponse"
  Unauthorized:
    description: "The credentials or token are invalid or expired."
    schema:
      $ref: "#/definitions/ErrorResponse"
...
  NotFound:
    description: "The requested resource was not found."
    schema:
      $ref: "#/definitions/ErrorResponse"
  TooManyRequests:
    description: "The request exceeded an applicable rate limit."
    schema:
      $ref: "#/definitions/ErrorResponse"
definitions:
...
  TicketPayload:
    type: "object"
    required:
      - "vin"
      - "ticket_date"
      - "ticket_desc"
      - "customer_id"
    properties:
      vin:
        type: "string"
        example: "ABCDEFGHIJKL1234567890"
      ticket_date:
        type: "string"
        format: "date"
        example: "2026-07-21"
      ticket_desc:
        type: "string"
        example: "Replace brake pads"
      customer_id:
        type: "integer"
        minimum: 1
        example: 1
  TicketResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        example: 1
      vin:
        type: "string"
        example: "ABCDEFGHIJKL1234567890"
      ticket_date:
        type: "string"
        format: "date"
        example: "2026-07-21"
      ticket_desc:
        type: "string"
        example: "Replace brake pads"
      customer_id:
        type: "integer"
        example: 1
      mechanic_ids:
        type: "array"
        items:
          type: "integer"
        example: [1, 2]
      inventory_items:
        type: "array"
        items:
          $ref: "#/definitions/TicketInventoryItem"
  TicketInventoryItem:
    type: "object"
    properties:
      inventory_id:
        type: "integer"
        example: 1
      name:
        type: "string"
        example: "Brake Pad Set"
      price:
        type: "number"
        format: "float"
        example: 59.99
      quantity:
        type: "integer"
        minimum: 1
        example: 2
  EditTicketMechanicsPayload:
    type: "object"
    required:
      - "add_ids"
      - "remove_ids"
    properties:
      add_ids:
        type: "array"
        description: "Mechanic IDs to add to the ticket."
        items:
          type: "integer"
        example: [2, 3]
      remove_ids:
        type: "array"
        description: "Mechanic IDs to remove from the ticket."
        items:
          type: "integer"
        example: [1]
  AddPartToTicketPayload:
    type: "object"
    properties:
      quantity:
        type: "integer"
        minimum: 1
        default: 1
        example: 2
...
  MessageResponse:
    type: "object"
    properties:
      message:
        type: "string"
  ErrorResponse:
    type: "object"
    description: "Error responses may use error or message depending on the handler. Validation errors may include field-specific properties."
    additionalProperties: true
    properties:
      error:
        type: "string"
      message:
        type: "string"